fix: minimizing extra files and cleaning up linter errors
