--- src/components/stepWizard/example usage/content.tsx
@@ -1,5 +1,0 @@
-const Page = () => {
-  return <div>Page Content</div>;
-};
-
-export default Page;
